Fix bug while trying to run main script
