--- src/dom.js
@@ -1,21 +1,0 @@
-/* eslint-disable import/no-cycle */
-import * as logic from './logic';
-/* eslint-enable import/no-cycle */
-
-const displayData = (data, units) => {
-  const giphy = document.getElementById('giphy');
-  const icon = document.getElementById('icon');
-  const description = document.getElementById('description');
-  description.textContent = data.weather[0].description;
-  logic.getGiphy(description, giphy);
-  icon.src = `https://openweathermap.org/img/wn/${data.weather[0].icon}@4x.png`;
-  const temp = document.getElementById('temp');
-  const wind = document.getElementById('wind');
-  const hum = document.getElementById('hum');
-  const pre = document.getElementById('pre');
-  [temp.textContent, wind.textContent, hum.textContent, pre.textContent] = logic.convertData(
-    data, units,
-  );
-};
-
-export default { displayData };
