+++ src/logic.js
@@ -1,16 +1,13 @@
-/* eslint-disable import/no-cycle */
-import * as dom from './dom';
-/* eslint-enable import/no-cycle */
 const forecast = async (place, units) => {
   await fetch(
     `https://api.openweathermap.org/data/2.5/weather?q=${place}&units=${units}&lang=en&appid=b863a920259d022cf48f60daf142113b`,
   ).then(
     (response) => response.json(),
-  ).then((res) => res).then((data) => dom.displayData(data, units));
+  ).then((res) => res).then((data) => displayData(data, units));
 };
 
 const convertData = (data, units) => {
-  let tempd = ((data.main.temp_max + data.main.temp_min) / 2);
+  let tempd = Math.round(((data.main.temp_max + data.main.temp_min) / 2)*100)/100;
   let windd = Math.round(data.wind.speed * 100) / 100;
   const humd = Math.round(data.main.humidity * 100) / 100;
   const pre = data.main.pressure;
@@ -35,5 +32,22 @@ const getGiphy = async (search, giphy) => {
     giphy.src = '#';
   }
 };
+/* eslint-disable no-use-before-define */
+const displayData = (data, units) => {
+  const giphy = document.getElementById('giphy');
+  const icon = document.getElementById('icon');
+  const description = document.getElementById('description');
+  description.textContent = data.weather[0].description;
+  getGiphy(description, giphy);
+  icon.src = `https://openweathermap.org/img/wn/${data.weather[0].icon}@4x.png`;
+  const temp = document.getElementById('temp');
+  const wind = document.getElementById('wind');
+  const hum = document.getElementById('hum');
+  const pre = document.getElementById('pre');
+  [temp.textContent, wind.textContent, hum.textContent, pre.textContent] = convertData(
+    data, units,
+  );
+};
+/* eslint-enable no-use-before-define */
 
-export { forecast, convertData, getGiphy };
+export { forecast };
